Move extended tests to separate Go packages
