--- test/extended/util.go
@@ -1,13 +1,0 @@
-package extended
-
-import (
-	"os"
-
-	"k8s.io/kubernetes/test/e2e"
-)
-
-var testContext e2e.TestContextType
-
-func kubeConfigPath() string {
-	return os.Getenv("KUBECONFIG")
-}
